Its best if I put these extension functions in an object so they aren't automatically exposed to our developers -- at least they have to import the methods.
